move symbol
